requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
celery==5.3.6
pydantic==2.5.3
pillow==10.1.0
requests==2.31.0
redis==5.0.1
